+++ CHANGELOG.md
@@ -4,6 +4,9 @@ v1.7.8
 Fixed:
 
 * Added a check for "Session Openness" when starting events. If the first event of a session is set to "No Join", Server Manager will correct it to "Free Join" (for Practice/Qualifying) and "Free join until 20 seconds to the Green Light" for Race sessions. This fixes an issue where the acServer will constantly loop sessions if the first session is not joinable.
+* Race Weekends now correctly handle entrants which have their car choice set to "Any Available Car"
+
+---
 
 v1.7.7
 ------